renamed ReflexTest to ReactionTest
