+++ AimReactionAPI/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost;Database=AimReactionDB;Trusted_Connection=True;Encrypt=False;"
+    "DefaultConnection": "Server=localhost,1433;Database=AimReactionDB;User Id=SA;Password=Eimantas1@;TrustServerCertificate=True;"
   }
 }
