+++ aim-reaction-app/src/pages/MovingTargets/MovingTargets.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 const MovingTarget: React.FC = () => {
     return (
       <div>
-        {/* You can add content or logic here later */}
+            MOVING TARGETS
       </div>
     );
   };
